+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/ChatUIInitializer.kt
@@ -17,19 +17,11 @@
 package io.getstream.chat.android.ui.common
 
 import android.content.Context
-import android.os.Build
 import androidx.startup.Initializer
-import coil.decode.GifDecoder
-import coil.decode.ImageDecoderDecoder
-import coil.decode.VideoFrameDecoder
-import com.getstream.sdk.chat.coil.StreamCoil
-import com.getstream.sdk.chat.coil.StreamImageLoaderFactory
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.header.VersionPrefixHeader
 import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import io.getstream.chat.android.ui.ChatUI
-import io.getstream.chat.android.ui.common.internal.AvatarFetcherFactory
-import io.getstream.chat.android.ui.common.internal.AvatarKeyer
 import kotlinx.coroutines.runBlocking
 
 /**
@@ -40,26 +32,6 @@ public class ChatUIInitializer : Initializer<Unit> {
     override fun create(context: Context): Unit = runBlocking(DispatcherProvider.IO) {
         ChatClient.VERSION_PREFIX_HEADER = VersionPrefixHeader.UI_COMPONENTS
         ChatUI.appContext = context
-
-        setImageLoader(context)
-    }
-
-    private fun setImageLoader(context: Context) {
-        val imageLoaderFactory = StreamImageLoaderFactory(context) {
-            components {
-                // duplicated as we can not extend component
-                // registry of existing image loader builder
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-                    add(ImageDecoderDecoder.Factory(enforceMinimumFrameDelay = true))
-                } else {
-                    add(GifDecoder.Factory(enforceMinimumFrameDelay = true))
-                }
-                add(AvatarFetcherFactory())
-                add(AvatarKeyer)
-                add(VideoFrameDecoder.Factory())
-            }
-        }
-        StreamCoil.setImageLoader(imageLoaderFactory)
     }
 
     override fun dependencies(): MutableList<Class<out Initializer<*>>> = mutableListOf()